# packages/contracts/openapi/control-plane.v1.yaml
openapi: 3.1.0
info:
  title: LaunchScope Control Plane API
  version: 1.0.0
  description: |
    Product-agnostic REST and SSE contract for the LaunchScope control plane.
    PostgreSQL is the business-state authority. This document defines transport
    semantics only; it does not select a Demo product or enable external calls.
  x-contract-version: '1.0'
  x-version-policy: independent-api-version
servers:
  - url: /api/v1
...
paths:
  /tenants:
    post:
      tags: [Identity]
      operationId: createTenant
      summary: Create a tenant and its owner workspace
      parameters:
        - $ref: '#/components/parameters/ActorId'
        - $ref: '#/components/parameters/IdempotencyKey'
        - $ref: '#/components/parameters/WriteCorrelationId'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [slug, workspace_name]
              properties:
                slug: {type: string, minLength: 1, maxLength: 120}
                workspace_name: {type: string, minLength: 1, maxLength: 200}
              additionalProperties: false
      responses:
        '201': {description: Tenant and workspace created}
        '400': {$ref: '#/components/responses/ValidationError'}
  /workspaces:
    post:
      tags: [Identity]
      operationId: createWorkspace
      summary: Create a tenant-local workspace for a tenant member
      parameters:
        - $ref: '#/components/parameters/TenantId'
        - $ref: '#/components/parameters/ActorId'
        - $ref: '#/components/parameters/IdempotencyKey'
        - $ref: '#/components/parameters/WriteCorrelationId'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [name]
              properties:
                name: {type: string, minLength: 1, maxLength: 200}
              additionalProperties: false
      responses:
        '201': {description: Workspace created}
        '403': {$ref: '#/components/responses/PolicyBlocked'}
...
components:
  parameters:
...
    TenantId:
      name: X-Tenant-Id
      in: header
      required: true
      schema:
        $ref: '#/components/schemas/Uuid'
    ActorId:
      name: X-Actor-Id
      in: header
      required: true
      schema:
        type: string
        minLength: 1
        maxLength: 255
...
    IdempotencyKey:
      name: Idempotency-Key
      in: header
      required: true
      description: 'Stable caller-generated key scoped to tenant, operation and request payload.'
      schema:
        type: string
        minLength: 1
        maxLength: 200
        pattern: '^[A-Za-z0-9][A-Za-z0-9._:-]{0,199}$'
    WriteCorrelationId:
      name: X-Correlation-Id
      in: header
      required: true
      description: 'Trace and causation grouping identifier. The response echoes it.'
      schema:
        $ref: '#/components/schemas/Uuid'
...
  headers:
    CorrelationId:
      description: 'Correlation identifier for this response and its server-side audit trail.'
      schema:
        $ref: '#/components/schemas/Uuid'
  responses:
...
    ValidationError:
      description: Request or cursor validation failed
      headers:
        X-Correlation-Id:
          $ref: '#/components/headers/CorrelationId'
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
...
    PolicyBlocked:
      description: Policy, budget or approval gate blocked the operation
      headers:
        X-Correlation-Id:
          $ref: '#/components/headers/CorrelationId'
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
...
  schemas:
    Uuid:
      type: string
      format: uuid
...
    ErrorResponse:
      type: object
      required: [error_code, message, correlation_id, retryable, details]
      properties:
        error_code:
          type: string
          enum:
            - VALIDATION_ERROR
            - UNAUTHORIZED
            - FORBIDDEN
            - NOT_FOUND
            - CONFLICT
            - IDEMPOTENCY_CONFLICT
            - PRECONDITION_FAILED
            - CURSOR_INVALID
            - RATE_LIMITED
            - DEPENDENCY_UNAVAILABLE
            - BUDGET_EXCEEDED
            - POLICY_BLOCKED
            - SUBMISSION_UNKNOWN
            - INTERNAL_ERROR
        message:
          type: string
          minLength: 1
        correlation_id:
          $ref: '#/components/schemas/Uuid'
        retryable:
          type: boolean
        details:
          type: object
      additionalProperties: false
...